Added getversion(), tested for -u -f w/jobid=None
In older versions of torque, qstat cannot take both -u and -f at
the same time. If a list of jobids is provided, then -u can be
ignored (and -a used if -f isn't used, to get the same output as
-u + jobids). If -u and -f are given without jobids, i.e. requesting
all the -f jobs of a given user, first qselect is used to get the
jobids, which are then passed as jobids to qstat -a or qstat -f
as appropriate.