Updates to make pbs compatible with slurm
software-specific functions (e.g. qstat) have been migrated
from misc to misc_torque, and  misc_slurm has been set up to
be a stand-in replacement for the functionality provided by
torque-specific commands. Syntax outside of misc_foo remains
the same, and all the publicly accessed methods look identical.

Import is performed at the class level, either auto-detecting
software or reading from a config file in the PBS_JOB_DB directory.

misc_torque._qstat has been made into a private function (and replaced
with misc_slurm._squeue) as this function is NOT identical between
programs, nor is it currently used outside of misc_foo (i.e., nothing
in either the Job or JobDB classes, or anything in the CASM python
code, directly calls misc.qstat())

This code is written for Slurm 2.3.3 or higher. If someone is feeling
ambitious, they can build-in version checking and on newer versions
of Slurm change from '-o %formats' to '-O formats' in the squeue
commands (the more explicit -O is not supported in 2.3.3, which is
all I have to work with)